Adding solution for problem 48

+++ src/main/java/be/vandemalle/jasper/project/euler/problems021to040/Problem028.java
@@ -31,7 +31,7 @@ public class Problem028 extends AbstractProblemTester {
 			}
 		}
 
-		return Long.toString(sumOfDiagonals);
+		return Integer.toString(sumOfDiagonals);
 	}
 
 	/** {@inheritDoc} */
